requirements.txt
fastapi>=0.111.0,<0.112.0
uvicorn[standard]>=0.30.0,<0.31.0
sqlmodel>=0.0.19,<0.0.20
sqlalchemy>=2.0.30,<2.1.0
pydantic>=2.8.0,<2.9.0
python-dotenv>=1.0.1,<2.0.0
httpx>=0.27.0,<0.28.0
tenacity>=8.3.0,<8.4.0
google-generativeai>=0.6.0,<0.7.0
pytest>=8.2.0,<8.3.0
pytest-asyncio>=0.23.0,<0.24.0
apscheduler>=3.10.4,<3.11.0

